+++ datacontract/engines/soda/check_soda_execute.py
@@ -1,6 +1,9 @@
 import logging
+import typing
 
-from pyspark.sql import SparkSession
+if typing.TYPE_CHECKING:
+    from pyspark.sql import SparkSession
+
 from soda.scan import Scan
 
 from datacontract.engines.soda.connections.bigquery import to_bigquery_soda_configuration
@@ -10,13 +13,14 @@ from datacontract.engines.soda.connections.kafka import create_spark_session, re
 from datacontract.engines.soda.connections.postgres import to_postgres_soda_configuration
 from datacontract.engines.soda.connections.snowflake import to_snowflake_soda_configuration
 from datacontract.engines.soda.connections.sqlserver import to_sqlserver_soda_configuration
+from datacontract.engines.soda.connections.trino import to_trino_soda_configuration
 from datacontract.export.sodacl_converter import to_sodacl_yaml
 from datacontract.model.data_contract_specification import DataContractSpecification, Server
 from datacontract.model.run import Run, Check, Log
 
 
 def check_soda_execute(
-    run: Run, data_contract: DataContractSpecification, server: Server, spark: SparkSession, tmp_dir
+    run: Run, data_contract: DataContractSpecification, server: Server, spark: 'SparkSession', tmp_dir
 ):
     if data_contract is None:
         run.log_warn("Cannot run engine soda-core, as data contract is invalid")
@@ -66,8 +70,10 @@ def check_soda_execute(
             scan.set_data_source_name(server.type)
     elif server.type == "dataframe":
         if spark is None:
-            run.log_warn("Server type dataframe only works with the Python library and requires a Spark session, "
-                         "please provide one with the DataContract class")
+            run.log_warn(
+                "Server type dataframe only works with the Python library and requires a Spark session, "
+                "please provide one with the DataContract class"
+            )
             return
         else:
             logging.info("Use Spark to connect to data source")
@@ -81,6 +87,10 @@ def check_soda_execute(
         scan.set_data_source_name(server.type)
     elif server.type == "sqlserver":
         soda_configuration_str = to_sqlserver_soda_configuration(server)
+        scan.add_configuration_yaml_str(soda_configuration_str)
+        scan.set_data_source_name(server.type)
+    elif server.type == "trino":
+        soda_configuration_str = to_trino_soda_configuration(server)
         scan.add_configuration_yaml_str(soda_configuration_str)
         scan.set_data_source_name(server.type)