+++ datacontract/export/odcs_converter.py
@@ -3,6 +3,12 @@ from typing import Dict
 import yaml
 
 from datacontract.model.data_contract_specification import DataContractSpecification, Model, Field
+from datacontract.export.exporter import Exporter
+
+
+class OdcsExporter(Exporter):
+    def export(self, data_contract, model, server, sql_server_type, export_args) -> dict:
+        return to_odcs_yaml(data_contract)
 
 
 def to_odcs_yaml(data_contract_spec: DataContractSpecification):
@@ -24,12 +30,29 @@ def to_odcs_yaml(data_contract_spec: DataContractSpecification):
 
     if data_contract_spec.terms is not None:
         odcs["description"] = {
-            "purpose": None,
+            "purpose": data_contract_spec.terms.description.strip()
+            if data_contract_spec.terms.description is not None
+            else None,
             "usage": data_contract_spec.terms.usage.strip() if data_contract_spec.terms.usage is not None else None,
             "limitations": data_contract_spec.terms.limitations.strip()
             if data_contract_spec.terms.limitations is not None
             else None,
         }
+
+    if data_contract_spec.servicelevels is not None:
+        slas = []
+        if data_contract_spec.servicelevels.availability is not None:
+            slas.append(
+                {
+                    "property": "generalAvailability",
+                    "value": data_contract_spec.servicelevels.availability.description,
+                }
+            )
+        if data_contract_spec.servicelevels.retention is not None:
+            slas.append({"property": "retention", "value": data_contract_spec.servicelevels.retention.period})
+
+        if len(slas) > 0:
+            odcs["slaProperties"] = slas
 
     odcs["type"] = "tables"  # required, TODO read from models.type?
     odcs["dataset"] = []